Add Dark mode, initialize header component, and start complexity info box
