+++ src/App.css
@@ -3,6 +3,9 @@
   margin: 0 auto;
   padding: 20px;
   text-align: center;
+  height: 100vh; /* Set the height to 100% of the viewport height */
+  flex-direction: column; /* Stack items vertically */
+  justify-content: center; /* Center content vertically */
 }
 
 textarea {
